im change veriables.

+++ src/Task_1_2_13/Okr.java
@@ -5,9 +5,9 @@ public class Okr {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("square-");
-        int area = in.nextInt();
+        double area = in.nextInt();
         double ros= calcCircleRadius( area);
-        System.out.printf("square-%d radius-%.3f",+area,ros);
+        System.out.printf("square-%.0f radius-%.3f",+area,ros);
